requirements.txt
aiokafka==0.7.2
anyio==3.5.0
asgiref==3.5.0
bidict==0.21.4
click==8.0.4
colorama==0.4.4
dynaconf==3.1.7
fastapi==0.75.0
h11==0.13.0
httptools==0.4.0
idna==3.3
kafka-python==2.0.2
numpy==1.22.3
opencv-python==4.5.5.64
pydantic==1.9.0
python-dotenv==0.19.2
python-engineio==4.3.1
python-socketio==5.5.2
PyYAML==6.0
sniffio==1.2.0
starlette==0.17.1
typer==0.4.0
typing_extensions==4.1.1
uvicorn==0.17.6
watchgod==0.8.1
websockets==10.2
